+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog for v0.0.5
 
-* add matrix chat
+* add matrix chat link to README
 * Display correct uv's while doing a face translate
 * Add option to compile with a http version checker
+* Set window title to map name on open. Fix segfault with sdl set window title.